# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/class/create": {
            "post": {
                "description": "教师创建班级，返回班级信息和邀请码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "class"
                ],
                "summary": "教师创建班级",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "创建班级参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/class.CreateClassRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/class.CreateClassResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/class/info/{class_id}": {
            "get": {
                "description": "获取班级基本信息，单参数使用 path",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "class"
                ],
                "summary": "获取班级详情",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "班级ID",
                        "name": "class_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/class.GetClassInfoResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/class/join": {
            "post": {
                "description": "学生使用邀请码加入班级",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "class"
                ],
                "summary": "学生加入班级",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "邀请码",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/class.JoinClassRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/class/my-created": {
            "get": {
                "description": "获取当前老师创建的所有班级",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "class"
                ],
                "summary": "获取教师创建的班级列表",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/class.GetMyCreatedClassesResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/class/my-joined": {
            "get": {
                "description": "获取当前学生加入的所有班级",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "class"
                ],
                "summary": "获取学生加入的班级列表",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/class.GetMyJoinedClassesResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/class/quit/{class_id}": {
            "post": {
                "description": "学生退出指定班级，单参数使用 path",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "class"
                ],
                "summary": "学生退出班级",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "班级ID",
                        "name": "class_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/class/students/{class_id}": {
            "get": {
                "description": "获取指定班级的所有学生，单参数使用 path",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "class"
                ],
                "summary": "获取班级内学生列表",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "班级ID",
                        "name": "class_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/class.GetClassStudentsResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "class.ClassInfo": {
            "type": "object",
            "properties": {
                "class_id": {
                    "type": "integer"
                },
                "class_name": {
                    "type": "string"
                },
                "teacher_id": {
                    "type": "integer"
                },
                "teacher_name": {
                    "type": "string"
                }
            }
        },
        "class.CreateClassRequest": {
            "type": "object",
            "required": [
                "class_name"
            ],
            "properties": {
                "class_name": {
                    "type": "string"
                }
            }
        },
        "class.CreateClassResp": {
            "type": "object",
            "properties": {
                "class_id": {
                    "type": "integer"
                },
                "class_name": {
                    "type": "string"
                },
                "invite_code": {
                    "type": "string"
                }
            }
        },
        "class.GetClassInfoResp": {
            "type": "object",
            "properties": {
                "class_info": {
                    "$ref": "#/definitions/class.ClassInfo"
                }
            }
        },
        "class.GetClassStudentsResp": {
            "type": "object",
            "properties": {
                "students": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/class.Student"
                    }
                }
            }
        },
        "class.GetMyCreatedClassesResp": {
            "type": "object",
            "properties": {
                "class_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/class.ClassInfo"
                    }
                }
            }
        },
        "class.GetMyJoinedClassesResp": {
            "type": "object",
            "properties": {
                "class_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/class.ClassInfo"
                    }
                }
            }
        },
        "class.JoinClassRequest": {
            "type": "object",
            "required": [
                "invite_code"
            ],
            "properties": {
                "invite_code": {
                    "type": "string"
                }
            }
        },
        "class.Student": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "http.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
    }
}